./2-alphabet_x10 | issue

+++ 0x02-functions_nested_loops/2-print_alphabet_x10.c
@@ -7,7 +7,7 @@
 */
 void print_alphabet_x10(void)
 {
-	short loop;
+	int loop;
 	char character = ('a' - 1);
 
 	for (loop = 0; (++loop) <= 10;)
